// packages/server/src/generated/api.ts
export type GetProjectAccessToken200 = {
  access_token: string;
  expires_in: number;
  token_type: string;
};

export type GetProjectAccessToken400 = {
  error: string;
};

export type VerifyAuthenticationTokenBody = {
  token: string;
};

export type VerifyAuthenticationToken200Payload = { [key: string]: unknown };

export type VerifyAuthenticationToken200 = {
  error: string;
  expiresAt: string;
  payload: VerifyAuthenticationToken200Payload;
  projectId: string;
  userId: string;
  valid: boolean;
};
// ...
export type BeginCredentialRegistration200Options = { [key: string]: unknown };

export type BeginCredentialRegistration200 = {
  authenticationToken: string;
  options: BeginCredentialRegistration200Options;
};

export type BeginCredentialRegistration400 = {
  error: string;
};

export type BeginCredentialRegistration401 = {
  error: string;
};

export type BeginCredentialRegistration500 = {
  error: string;
};

export type GetUser200 = {
  id: string;
  userName?: string;
};
// ...
export type CreateUser200 = {
  id: string;
  userName?: string;
};

export type CreateUser400 = {
  error: string;
};

export type CreateUser500 = {
  error: string;
};

export type UpdateUserBodyUpdates = {
  userName?: string | null;
};

export type UpdateUserBodyUserIdentifier = {
  userId?: string;
  userName?: string;
};

export type UpdateUserBody = {
  updates: UpdateUserBodyUpdates;
  userIdentifier: UpdateUserBodyUserIdentifier;
};

export type UpdateUser200 = {
  id: string;
  userName?: string;
};
// ...
export type BulkCreateUsers200Item = {
  id: string;
  userName?: string;
};

export type BulkCreateUsers400 = {
  error: string;
};

export type GetUserCredentials200Item = {
  aaguid: string | null;
  authenticatorType: string | null;
  counter: number | null;
  createdAt: string;
  credentialBackedUp: boolean | null;
  credentialDeviceType: string | null;
  id: string;
  label: string | null;
  lastUsedAt: string | null;
  transports: string[] | null;
  updatedAt: string | null;
  userId: string;
  webAuthnId: string | null;
};

export type GetUserCredentials401 = {
  error: string;
};

export type GetUserCredentials404 = {
  error: string;
};

export type GetCredential200 = {
  aaguid: string | null;
  authenticatorType: string | null;
  counter: number | null;
  createdAt: string;
  credentialBackedUp: boolean | null;
  credentialDeviceType: string | null;
  id: string;
  label: string | null;
  lastUsedAt: string | null;
  transports: string[] | null;
  updatedAt: string | null;
  userId: string;
  webAuthnId: string | null;
};